expandable cards exercise finished
