+++ 001_expandable_cards/README.md
@@ -1,0 +1,22 @@
+# Day 001 Expandable Cards
+
+### Demo:
+<image src="./gif/001_expandable_card.gif"/>
+
+### Knowledges:
+This exercise is implemented with basic React concepts like Dynamic Values, Props, and Hook. It is a handy and beginner-friendly exercise that can help get used to React expressions by integrating content-interaction-visualization in one simple widget. 
+
+### How to run the file:
+Open the terminal and navigate to this folder, then excuting the following commands to install the necessary packages:
+```bash
+npm install
+```
+Then use 
+```bash
+npm run dev
+``` 
+to activate the local server, paste the link showing in termial that starts with ``http://localhost:****/`` (where `****` is a 4-digit port number that varies between machines) to the browser.
+
+### Provisional Use Case:
+- Introducing team members
+- Displaying major categories of products/services
