+++ 001_expandable_cards/src/App.css
@@ -1,0 +1,75 @@
+@import url("https://fonts.googleapis.com/css2?family=Quicksand:wght@400;700&family=Roboto+Condensed:wght@400;700&display=swap");
+
+body {
+  font-family: "Quicksand", sans-serif;
+  margin: 0;
+  display: flex;
+  place-items: center;
+  min-width: 320px;
+  min-height: 100vh;
+}
+
+@media (prefers-color-scheme: light) {
+  :root {
+    color: #213547;
+    background-color: #ffffff;
+  }
+}
+
+#root {
+  max-width: 1280px;
+  height: 95vh;
+  width: 95vw;
+  margin: auto;
+  padding: 1rem;
+  text-align: center;
+  background: lightgray;
+  box-sizing: border-box;
+  border-radius: 25px
+}
+
+h1{
+  margin: 0.5em;
+  font-size: 30px;
+  font-weight: bold;
+}
+
+h2{
+  margin: 0.4em;
+  font-size: 22px;
+  font-weight: bold;
+}
+
+p{
+  margin: 0.2em;
+}
+
+section{
+  align-content:center;
+}
+
+.container{
+  margin: 30px auto 20px auto;
+  position: relative;
+  display:grid;
+  width: 70%;
+  height: 55vh;
+  gap: 1em;
+  transition: 300ms;
+}
+
+.info {
+    background: white;
+    opacity: 1.0;
+    margin: auto;
+    margin-top: 40px;
+    display: block;
+    position: relative;
+    width: 55%;
+    padding: 5px 10px 20px 10px;
+    border-radius: 20px;
+}
+
+.info p{
+  font-size: 15px;
+}
