+++ 001_expandable_cards/src/components/MemberCard.css
@@ -1,0 +1,11 @@
+.box{
+    border-radius: 10px;
+  }
+
+.box img{
+    width:100%;
+    height:100%;
+    object-fit:cover;
+    transition: 400ms;
+    border-radius: 10px;
+}
