--- 001_team_members/README.md
@@ -1,9 +1,0 @@
-# Day 001 Team Members
-
-Packages:
-- Vite
-- React
-- Eslint
-
-`npm install`
-`npm run dev`
