--- 001_team_members/src/App.css
@@ -1,69 +1,0 @@
-#root {
-  max-width: 1280px;
-  height: 95vh;
-  width: 95vw;
-  margin: auto;
-  padding: 1rem;
-  text-align: center;
-  background: lightgray;
-  box-sizing: border-box;
-  border-radius: 25px
-}
-
-h1{
-  margin: 0.5em;
-  font-size: 30px;
-  font-weight: bold;
-}
-
-h2{
-  margin: 0.4em;
-  font-size: 22px;
-  font-weight: bold;
-}
-
-p{
-  margin: 0.2em;
-}
-
-section{
-  aligh-content:center;
-}
-
-.container{
-  margin: 30px auto 20px auto;
-  position: relative;
-  display:grid;
-  width: 70%;
-  height: 55vh;
-  gap: 1em;
-  transition: 300ms;
-}
-
-.box{
-    border-radius: 10px;
-  }
-
-.box img{
-    width:100%;
-    height:100%;
-    object-fit:cover;
-    transition: 400ms;
-    border-radius: 10px;
-}
-
-.info {
-    background: white;
-    opacity: 1.0;
-    margin: auto;
-    margin-top: 40px;
-    display: block;
-    position: relative;
-    width: 55%;
-    padding: 5px 10px 20px 10px;
-    border-radius: 20px;
-}
-
-.info p{
-  font-size: 15px;
-}
